requirements.txt
moviepy==1.0.3
srt_equalizer
termcolor
Pillow>=10.0.0
selenium
webdriver_manager